pyproject.toml
[project]
name = "safari_sdk"
description = "Safari SDK: the SDK for Google DeepMind Gemini Robotics models"
readme = "README.md"
requires-python = ">=3.10,<3.14"  # TODO
license = {file = "LICENSE"}
authors = [{name = "Safari SDKAuthors", email="safari-sdk-authors@google.com"}]
classifiers = [  # List of https://pypi.org/classifiers/
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",  # EOL 2026-10
    "Programming Language :: Python :: 3.11",  # EOL 2027-10
    "Programming Language :: Python :: 3.12",  # EOL 2028-10
    # "Programming Language :: Python :: 3.13", # EOL 2029-19, no compatible tensorflow version as of TF 2.18
    "License :: Other/Proprietary License",
    "Intended Audience :: Science/Research",
]
keywords = []

# pip dependencies of the project
dependencies = [
  # copybara:strip_begin(for internal changes only)
  # LINT.IfChange
  # copybara:strip_end
  # go/keep-sorted start
  "absl-py",
  "dataclasses_json",
  "dm-env",
  "evdev",
  "fastapi",
  "gdm-robotics",
  "google-api-python-client",
  "google-auth-httplib2",
  "google-auth-oauthlib",
  "google-genai<=1.55.0",
  "grpcio",
  "imageio",
  "immutabledict",
  "lark~=1.2",
  "mcap-protobuf-support",
  "mediapy",
  "mujoco>=3.3.6",
  "opencv-python",
  "overrides",
  "protobuf<7",
  "psutil",
  "python-magic",
  "pytz",
  "scipy",
  "tensorflow",
  "uvicorn",
  "watchdog",
  # go/keep-sorted end
  # copybara:strip_begin(for internal changes only)
  # LINT.ThenChange(//depot/google3/third_party/safari/sdk/requirements.txt)
  # Run `bash third_party/safari/sdk/scripts/check_dependency_version_requirements.sh`
  # to update the dependencies in requirements.txt.
  # copybara:strip_end
]

# Configure scikit to pull version value from "safari_sdk/__init__.py"
# See: https://scikit-build-core.readthedocs.io/en/latest/configuration/dynamic.html#regex
dynamic = ["version"]

[tool.scikit-build.metadata.version]
provider = "scikit_build_core.metadata.regex"
input = "safari_sdk/__init__.py"

[tool.scikit-build.wheel]
# By default, scikit-build assumes the package contains platform-specific files
# and names the .whl file accordingly (eg "*-cp313-cp313-linux_x86_64.whl").
# However, this package does not contain any platform-specific files, so set
# platlib to false so the output whl ends with "-py3-none-any.whl". If this ever
# changes (e.g. if we add pybind11 extensions), this flag should be removed.
platlib = false

[project.scripts]
flywheel-cli = "safari_sdk.flywheel.flywheel_cli:cli_main"  # 'flywheel-cli' will be installed as a command

[project.urls]
homepage = "https://deepmind.google/models/gemini-robotics/"
repository = "https://github.com/google-deepmind/gemini-robotics-sdk"
# Other: `documentation`, `changelog`

[project.optional-dependencies]
# Development deps (unittest, linting, formating,...)
# Installed through `pip install .[dev]`
dev = [
    # go/keep-sorted start
    "immutabledict",
    "parameterized",
    "pyink",
    "pylint>=2.6.0",
    "pytest",
    "pytest-asyncio",
    "pytest-xdist",
    # go/keep-sorted end
]


[tool.pyink]
# Formatting configuration to follow Google style-guide
pyink-indentation = 2
pyink-use-majority-quotes = true

# The scikit backend invokes cmake to build the proto messages.
# This is done in a unique temporary build directory unless otherwise specified.
[build-system]
requires = ["scikit-build-core>=0.10","setuptools"]
build-backend = "scikit_build_core.build"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
python_files = ["*_test.py"]
# Use asyncio for async tests (eg safari_sdk/agent/framework/event_bus/event_stream_handler_test.py).
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
